[lab05] Created

+++ src/com/alexey/minay/labs/lab04/shape/shapes/LineSegment.kt
@@ -10,7 +10,7 @@ class LineSegment : Shape {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getOutlineColor(): Double {
+    override fun getOutlineColor(): Int {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
